requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==5.0.0
SQLAlchemy==2.0.45
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0
jinja2==3.1.6
aiofiles==25.1.0
requests==2.32.3
httpx==0.28.1
openai==2.15.0
langchain==1.2.6
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
pytz==2025.2
pillow==12.1.0
orjson==3.11.5
